requirements.txt
azure-core==1.32.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
idna==3.10
isodate==0.7.2
lorem-text==2.1
pycparser==2.22
requests==2.32.3
six==1.16.0
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
